Extract library view mappings
