+++ src/apps/experimental/routes/music/index.tsx
@@ -1,84 +1,14 @@
-import { BaseItemKind } from '@jellyfin/sdk/lib/generated-client/models/base-item-kind';
-import React, { FC } from 'react';
-import useCurrentTab from 'hooks/useCurrentTab';
-import Page from 'components/Page';
-import PageTabContent from '../../components/library/PageTabContent';
-import { LibraryTab } from 'types/libraryTab';
 import { CollectionType } from '@jellyfin/sdk/lib/generated-client/models/collection-type';
-import { LibraryTabContent, LibraryTabMapping } from 'types/libraryTabContent';
-import { MusicSuggestionsSectionsView } from 'types/sections';
+import React, { type FC } from 'react';
 
-const albumArtistsTabContent: LibraryTabContent = {
-    viewType: LibraryTab.AlbumArtists,
-    collectionType: CollectionType.Music,
-    isBtnSortEnabled: false
-};
-
-const albumsTabContent: LibraryTabContent = {
-    viewType: LibraryTab.Albums,
-    collectionType: CollectionType.Music,
-    isBtnPlayAllEnabled: true,
-    isBtnShuffleEnabled: true,
-    itemType: [BaseItemKind.MusicAlbum]
-};
-
-const artistsTabContent: LibraryTabContent = {
-    viewType: LibraryTab.Artists,
-    collectionType: CollectionType.Music,
-    isBtnSortEnabled: false
-};
-
-const playlistsTabContent: LibraryTabContent = {
-    viewType: LibraryTab.Playlists,
-    isBtnFilterEnabled: false,
-    isBtnGridListEnabled: false,
-    isBtnNewPlaylistEnabled: true,
-    isAlphabetPickerEnabled: false,
-    itemType: [BaseItemKind.Playlist]
-};
-
-const songsTabContent: LibraryTabContent = {
-    viewType: LibraryTab.Songs,
-    isBtnShuffleEnabled: true,
-    isBtnGridListEnabled: false,
-    isAlphabetPickerEnabled: false,
-    itemType: [BaseItemKind.Audio]
-};
-
-const suggestionsTabContent: LibraryTabContent = {
-    viewType: LibraryTab.Suggestions,
-    collectionType: CollectionType.Music,
-    sectionsView: MusicSuggestionsSectionsView
-};
-
-const genresTabContent: LibraryTabContent = {
-    viewType: LibraryTab.Genres,
-    collectionType: CollectionType.Music,
-    itemType: [BaseItemKind.MusicAlbum]
-};
-
-const collectionsTabContent: LibraryTabContent = {
-    viewType: LibraryTab.Collections,
-    collectionType: CollectionType.Music,
-    isBtnNewCollectionEnabled: true,
-    itemType: [BaseItemKind.BoxSet],
-    noItemsMessage: 'MessageNoCollectionsAvailable'
-};
-
-const musicTabMapping: LibraryTabMapping = {
-    0: albumsTabContent,
-    1: suggestionsTabContent,
-    2: albumArtistsTabContent,
-    3: artistsTabContent,
-    4: playlistsTabContent,
-    5: songsTabContent,
-    6: genresTabContent,
-    7: collectionsTabContent
-};
+import PageTabContent from 'apps/experimental/components/library/PageTabContent';
+import viewsByKind from 'apps/experimental/features/libraries/constants/views';
+import Page from 'components/Page';
+import useCurrentTab from 'hooks/useCurrentTab';
 
 const Music: FC = () => {
     const { libraryId, activeTab } = useCurrentTab();
-    const currentTab = musicTabMapping[activeTab];
+    const currentTab = viewsByKind[CollectionType.Music][activeTab];
 
     return (
         <Page
